+++ Backend/DevopsTools/Dev/scripts/helper/env_config.sh
@@ -2,10 +2,11 @@ export NAMESPACE="leduynhan1201"
 export REPOSITORY_NAME="learn-dot-net"
 
 export ENVOY_TAG=tools-dev # https://hub.docker.com/r/envoyproxy/envoy/tags
+export BACKEND_TAG="1.0.0"
 
 LOCAL_IP=$(hostname -I | awk '{print $1}')
 export LOCAL_IP
-export GATEWAY_PORT=8889
+export GATEWAY_PORT=60000
 export BACKEND_HOSTNAME="backend.${REPOSITORY_NAME}.${MODE}"
 export CA_NAME="LDNhanRootCA"
 export SUBJ_C="VN"